+++ python/xorq/backends/let/tests/snapshots/test_cache/test_udaf_caching/py310_test_udaf_caching.txt
@@ -1,1 +1,1 @@
-letsql_cache-letsql_cache-7103b0af461289735304821523106b58
+letsql_cache-letsql_cache-9c661c21820ae0223618e4c64bf00f95
