+++ python/xorq/backends/let/tests/snapshots/test_cache/test_udaf_caching/py311_test_udaf_caching.txt
@@ -1,1 +1,1 @@
-letsql_cache-letsql_cache-a5017faee6594c3f71d4d2102513b68b
+letsql_cache-letsql_cache-e5cc4f3ff1e70aefdc4d8a17b5f5cba7
