+++ python/xorq/backends/let/tests/snapshots/test_cache/test_udaf_caching/py312_test_udaf_caching.txt
@@ -1,1 +1,1 @@
-letsql_cache-letsql_cache-2609daaddfbc2558aa8d82ffa23eeb83
+letsql_cache-letsql_cache-bc9b9adc0b836cbd0ee9c5c9d69b4f3c
