+++ python/xorq/backends/let/tests/snapshots/test_cache/test_udf_caching/py311_udf_caching.txt
@@ -1,1 +1,1 @@
-letsql_cache-letsql_cache-812433b4f1764359fda67dbc6a1720a3
+letsql_cache-letsql_cache-3e186bb7eb8710440e22b8004a34cf26
