+++ python/xorq/backends/let/tests/snapshots/test_cache/test_udf_caching/py312_udf_caching.txt
@@ -1,1 +1,1 @@
-letsql_cache-letsql_cache-00bcfd0fda67d72cf36feb9683a39dd6
+letsql_cache-letsql_cache-84af6947f5f3386df69014b3d27bcb6f
