+++ python/xorq/common/utils/tests/snapshots/test_dask_normalize/test_duckdb_snapshot_key/duckdb_snapshot_key.txt
@@ -1,1 +1,1 @@
-letsql_cache-snapshot-ca946f04872afce2de52e286be7c7e1f
+letsql_cache-snapshot-989e41aae729716526ca286edc8a135f
